Add support for building a shared library
Make it possible to build a shared library in addition to the static
one. As per policy of many distributions, static libraries for code
that is not internal to a project is frowned upon (for good reasons),
and open-iscsi does link against open-isns. Hence provide a shared
library for open-iscsi (and other proejcts) to link against.

This does not change the default build's behavior in only building and
installing a static library; when passing --enable-shared to configure
a shared library will be built and installed.

The shared library is now built with symbol versioning if available,
with all symbols not exported in public headers being either hidden
or marked with an internal version.

Signed-off-by: Christian Seiler <christian@iwakd.de>