Support FreeBSD and GNU/kFreeBSD
Some changes are required to make open-isns work on FreeBSD, mainly
related to the passing of credentials via AF_LOCAL sockets. In order
to support FreeBSD (and GNU/kFreeBSD), ad #ifdef checks for specific
fetatures (such as the availability of SCM_CREDENTIALS or SO_PASSCRED)
to the code.

The following changes are contained in this commit:

 - time_t requires #include <sys/time.h> with FreeBSD's libc
 - SO_PASSCRED is unsupported on FreeBSD, in order to receive
   credentials via recvmsg(), the sender needs to prepare an (empty)
   control message of the right type and use sendmsg()
 - <endian.h> doesn't exist for htonll/ntohll emulation, use
   <sys/endian.h> instead
 - FreeBSD's libc doesn't define s6_addr32 for userland code, but only
   s6_addr; add the corresponding define if it's missing
 - FreeBSD defaults to IPV6_V6ONLY = 1 (unlike Linux), set it to 0 for
   sockets explicitly (otherwise isnsd will only be reachable via IPv6
   by default)
 - instead of SCM_CREDENTIALS + struct ucred FreeBSD uses SCM_CREDS and
   struct cmsgcred with slightly different field names

Tested on both Debian GNU/kFreeBSD and upstream FreeBSD. (Kernel 10.3
in both cases.) Also tested on Debian GNU/Linux against regressions.

Signed-of-by: Christian Seiler <christian@iwakd.de>