DD: handle Domain ID change on creation
tests/test02.pl currently fails because when the Discovery Domain is
created, it's dd_id is 0. This ID is used to mark membership since
d6004b73 ('Fix DD member doubling when restoring from DB'). However,
isns_dd_insert takes a DD ID of 0 as a special case and sets it to the
next free ID (isns_dd_next_id), which starts at 1. When we iterate
through and append the temp_dd members to the actual dd object, we don't
reset the membership to be the actual DD's ID.

I noticed this fix by having test02 fail due to not seeing other members
of the DD, stopping and starting isnsd (but not deleting the database)
and then having the DD queries succeed. With this change, test02
succeeds.

This partially reverts d6004b73 ('Fix DD member doubling when restoring
from DB').