tests/data/test06: sort database responses
test06 is the only test that doesn't seem to have its responses sorted.
harness.pl does a side-by-side comparison of the actual and expected
responses, and there is no assertion that the database output will
always be in the same order (and given its keyed by indices, etc., it's
not sensible to demand this, as far as I can tell). Sort the expected
output.