tests/test06.pl: :iscsi suffix is not valid
As far as I can tell, the :iscsi suffix is not well-defined for portals
in iSNS. This leads to failures in test06 and I believe the intention is
to use port 860. Hard code this for now, until a maintainer determines
what should actually be done. There is some reference to ':isns' in
isnsdd.c and etc/isns{adm,dd}.conf, but it's not clear that it is
correctly (or actually) implemented.