Makefile: Fix shared object dependency handling
Use a pattern rule to match solib targets.

The existing rule tries to use target variable ($@) while specifying the
prerequisites. This is not valid (unless you use $$@ and
.SECONDEXPANSION, which relies on make > 3.8).

This fixes rebuilding with the following scenario:

./configure --enable-shared
make
touch server.c
make

Currently you would see the .a static library rebuild as the .o targets
for the non solib- prefixed targets relies on the regular .o target,
which has valid prerequisites. The solib- prefixed targets rely on the
solib-specific target which has the incorrect prerequisits, which means
that it sees an empty prerequisit, and doesnt rebuild the shared object
version.

Signed-off-by: Robert Beckett <bbeckett@netvu.org.uk>