Fix compiler complaint about possible alignment issue
The compiler was complaining when we took the address of
a structure member, where the structure was defined as
'packed', since doing so can result in the address not being
aligned on a 32-bit boundary. The fix is to take the address
of a stack variable, instead. The fix isn't beautiful,
but it is functional and easy to understand.