Fix broken server authentication initialization.
Commit 86bf736873ed cleaned up a bunch of compiler complaints,
and was supposed to have no effect on functionality. But the change
to isns_dsa_param_gen_callback() from void to returning an
integer (zero) broke "isnsd --init", causing the error message:

> Generating DSA parameters; this may take a while
> +
> Warning: Error generating DSA parameters - OpenSSL errors follow:

This commit changes isns_dsa_param_gen_callback() back to a void
function, and fixes the compiler problem by properly declaring
the callback, and using the proper openssl macro to set the
callback.

Fixes: 86bf736873ed General cleanup for the compiler.